Resolve some problem to execute sqlite, working on

+++ database/migrations/2020_08_08_123543_drop_product_id_column_into_invoice_products_table.php
@@ -14,6 +14,13 @@ class DropProductIdColumnIntoInvoiceProductsTable extends Migration
     public function up()
     {
         Schema::table('invoice_products', function (Blueprint $table) {
+            $table->dropColumn('product_id');
+        });
+
+        if (config('database.default') === 'sqlite') {
+            return;
+        }
+        Schema::table('invoice_products', function (Blueprint $table) {
             $table->dropForeign('invoice_products_product_id_foreign');
             $table->dropIndex('invoice_products_product_id_foreign');
             $table->dropColumn('product_id');
@@ -27,6 +34,16 @@ class DropProductIdColumnIntoInvoiceProductsTable extends Migration
      */
     public function down()
     {
+        if (config('database.default') === 'sqlite') {
+            Schema::table('invoice_products', function (Blueprint $table) {
+                $table->unsignedBigInteger('product_id')
+                    ->default(0);
+                $table->foreign('product_id')
+                    ->references('id')
+                    ->on('products');
+            });
+            return;
+        }
         Schema::table('invoice_products', function (Blueprint $table) {
             $table->unsignedBigInteger('product_id');
             $table->foreign('product_id')
